refactor: split JsonWebAPI into smaller files
